+++ tests/tester.nim
@@ -171,3 +171,80 @@ when defined(linux):
       check hasLine(output.processOutput, "switched to nim 1.0.0")
 
       check not dirExists(choosenimDir / "toolchains" / "nim-1.0.0" / "c_code")
+
+test "can update devel with git":
+  beginTest()
+  block:
+    let (output, exitCode) = exec(@["devel", "--latest"], liveOutput=true)
+    check exitCode == QuitSuccess
+
+    check inLines(output.processOutput, "extracting")
+    check inLines(output.processOutput, "setting")
+    check inLines(output.processOutput, "latest changes")
+    check inLines(output.processOutput, "building")
+
+  block:
+    let (output, exitCode) = exec(@["update", "devel", "--latest"], liveOutput=true)
+    check exitCode == QuitSuccess
+
+    check not inLines(output.processOutput, "extracting")
+    check not inLines(output.processOutput, "setting")
+    check inLines(output.processOutput, "updating")
+    check inLines(output.processOutput, "latest changes")
+    check inLines(output.processOutput, "building")
+
+test "can install and update nightlies":
+  beginTest()
+  block:
+    # Install nightly
+    let (output, exitCode) = exec("devel", liveOutput=true)
+
+    # Travis runs into Github API limit
+    if not inLines(output.processOutput, "unavailable"):
+      check exitCode == QuitSuccess
+
+      check inLines(output.processOutput, "devel from")
+      check inLines(output.processOutput, "setting")
+      when not defined(macosx):
+        if not inLines(output.processOutput, "recent nightly"):
+          check inLines(output.processOutput, "already built")
+      check inLines(output.processOutput, "to Nim #devel")
+
+      block:
+        # Update nightly
+        let (output, exitCode) = exec(@["update", "devel"], liveOutput=true)
+
+        # Travis runs into Github API limit
+        if not inLines(output.processOutput, "unavailable"):
+          check exitCode == QuitSuccess
+
+          check inLines(output.processOutput, "updating")
+          check inLines(output.processOutput, "devel from")
+          check inLines(output.processOutput, "setting")
+          when not defined(macosx):
+            if not inLines(output.processOutput, "recent nightly"):
+              check inLines(output.processOutput, "already built")
+
+      block:
+        # Update to devel latest
+        let (output, exitCode) = exec(@["update", "devel", "--latest"], liveOutput=true)
+        check exitCode == QuitSuccess
+
+        when not defined(macosx):
+          check not inLines(output.processOutput, "extracting")
+        check not inLines(output.processOutput, "setting")
+        check inLines(output.processOutput, "updating")
+        check inLines(output.processOutput, "latest changes")
+        check inLines(output.processOutput, "building")
+
+test "can update self":
+  # updateSelf() doesn't use options --choosenimDir and --nimbleDir. It's used getAppDir().
+  # This will rewrite $project/bin dir, it's dangerous.
+  # So, this test copy bin/choosenim to test/choosenimDir/choosenim, and use it.
+  beginTest()
+  let testExePath = choosenimDir / extractFilename(exePath)
+  copyFileWithPermissions(exePath, testExePath)
+  block :
+    let (output, exitCode) = exec(["update", "self", "--debug", "--force"], exe=testExePath, liveOutput=true)
+    check exitCode == QuitSuccess
+    check inLines(output.processOutput, "Info: Updated choosenim to version")